chore: move dashboard native UI updates to correct monorepo path apps/McomMall
