+++ apps/McomMall/app/dashboard/component/ActivityTimerBadge.tsx
