+++ apps/McomMall/app/dashboard/component/SideMenu.tsx
@@ -2,10 +2,14 @@
 
 import { MenuContent } from './MenuContent';
 
-const SideMenu = () => {
+interface SideMenuProps {
+  isCollapsed?: boolean;
+}
+
+const SideMenu = ({ isCollapsed }: SideMenuProps) => {
   return (
-    <aside className="w-full  p-4 bg-gray-100 rounded-2xl hide-scrollbar">
-      <MenuContent />
+    <aside className={`w-full p-4 bg-gray-100 rounded-2xl hide-scrollbar transition-all duration-300 ${isCollapsed ? 'items-center' : ''}`}>
+      <MenuContent isCollapsed={isCollapsed} />
     </aside>
   );
 };
