+++ apps/McomMall/app/dashboard/page.tsx
@@ -42,15 +42,10 @@ import RecentActivities from './component/RecentActivities';
 import EarningProgressionChart from './component/EarningProgressionChart';
 import { CustomerStatsDto, OwnerStatsDto } from '@/service/stats/types';
 import { UserRole } from '@/service/auth/types';
-import { Progress } from '@/components/ui/progress';
-import { useGetTiers } from '@/service/tiers/hook';
-import { ActivityTimerCard } from '@/components/dashboard/ActivityTimerCard';
-import { useGetActivityTimerStatus, usePauseActivityTimer, useResumeActivityTimer } from '@/service/activity-timer/hook';
-import { FloatingTrialTimer } from '@/components/dashboard/FloatingTrialTimer';
-import { ActivityTimerType } from '@/service/activity-timer/types';
-
+import { MobileDashboardHub } from './component/MobileDashboardHub';
 
 // --- TYPE DEFINITIONS ---
+
 interface ListingPackage {
   name: string;
   description: string;
@@ -67,135 +62,63 @@ const listingPackage: ListingPackage = {
 };
 
 // --- SUB-COMPONENTS ---
-const ListingPackages: FC<{
-  name: string;
-  current: number;
-  max: number;
-}> = ({ name, current, max }) => {
-  const percentage = Math.min(Math.round((current / max) * 100), 100);
-
-  return (
-    <Card className="shadow-sm">
-      <CardHeader>
-        <CardTitle className="text-lg font-semibold">
-          Your Listing Packages
-        </CardTitle>
-      </CardHeader>
-      <CardContent className="space-y-4">
-        <div className="flex items-start space-x-4">
-          <div className="flex-shrink-0 w-10 h-10 rounded-full bg-orange-100 flex items-center justify-center">
-            <Diamond className="w-5 h-5 text-orange-600" />
-          </div>
-          <div className="flex-grow">
-            <div className="flex justify-between items-center mb-1">
-              <p className="font-semibold text-gray-800">{name}</p>
-              <span className="text-xs font-medium text-gray-500">
-                {current} / {max}
-              </span>
-            </div>
-            <Progress value={percentage} className="h-2" />
-            <p className="text-sm text-gray-500 mt-2">
-              You have used {current} out of {max} listing slots.
-            </p>
-          </div>
-        </div>
-      </CardContent>
-    </Card>
-  );
-};
+const ListingPackages: FC<{ pkg: ListingPackage }> = ({ pkg }) => (
+  <Card className="shadow-sm">
+    <CardHeader>
+      <CardTitle className="text-lg font-semibold">
+        Your Listing Packages
+      </CardTitle>
+    </CardHeader>
+    <CardContent className="flex items-start space-x-4">
+      <div className="flex-shrink-0 w-10 h-10 rounded-full bg-gray-100 flex items-center justify-center">
+        <Diamond className="w-5 h-5 text-gray-600" />
+      </div>
+      <div>
+        <p className="font-semibold text-gray-800">{pkg.name}</p>
+        <p className="text-sm text-gray-500">{pkg.description}</p>
+      </div>
+    </CardContent>
+  </Card>
+);
 
 // --- MAIN PAGE COMPONENT ---
 const DashboardPage: FC = () => {
-  const { userName, userRole, packageInfo } = useSelector(
-    (state: RootState) => state.auth
-  );
-  const { data: stats, isLoading: isLoadingStats } = useGetStats<
-    OwnerStatsDto | CustomerStatsDto
-  >();
-  const { data: activities, isLoading: isLoadingActivities } =
-    useRecentActivities();
-  const { data: tiers } = useGetTiers();
-
-  const { data: timers, isLoading: isLoadingTimers } = useGetActivityTimerStatus();
-  const pauseTimerMutation = usePauseActivityTimer();
-  const resumeTimerMutation = useResumeActivityTimer();
-
-  const currentTier = tiers?.find(
-    t => t.name.toLowerCase() === packageInfo?.planType?.toLowerCase()
-  );
-
-  const maxListings = currentTier?.configuration?.quotas?.maxListings || 1;
-  const currentListings =
-    userRole === UserRole.OWNER
-      ? (stats as OwnerStatsDto)?.totalAmountOfListing || 0
-      : 0;
-
-  const isTrial = packageInfo?.planType?.toLowerCase() === 'trial';
-  const trialTimer = timers?.find(t => t.type === ActivityTimerType.TRIAL);
+  const { userName, userRole } = useSelector((state: RootState) => state.auth);
+  const { data: stats, isLoading: isLoadingStats } = useGetStats<OwnerStatsDto | CustomerStatsDto>();
+  const {
+    data: activities,
+    isLoading: isLoadingActivities,
+  } = useRecentActivities();
 
   return (
     <div className="min-h-screen bg-gray-50 p-4 sm:p-6 lg:p-8">
       {/* Header */}
       <header className="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-8">
-        <h1 className="text-3xl font-bold text-gray-800">
-          Hello {userName || 'User'} !
+        <h1 className="text-2xl sm:text-3xl font-bold text-gray-800">
+          Hello {userName || 'User'}!
         </h1>
-        <Breadcrumb className="mt-2 sm:mt-0">
-          <BreadcrumbList>
-            <BreadcrumbItem>
-              <BreadcrumbLink href="/">Home</BreadcrumbLink>
-            </BreadcrumbItem>
-            <BreadcrumbSeparator />
-            <BreadcrumbItem>
-              <BreadcrumbPage>Dashboard</BreadcrumbPage>
-            </BreadcrumbItem>
-          </BreadcrumbList>
-        </Breadcrumb>
+        <div className="hidden sm:block">
+          <Breadcrumb>
+            <BreadcrumbList>
+              <BreadcrumbItem>
+                <BreadcrumbLink href="/">Home</BreadcrumbLink>
+              </BreadcrumbItem>
+              <BreadcrumbSeparator />
+              <BreadcrumbItem>
+                <BreadcrumbPage>Dashboard</BreadcrumbPage>
+              </BreadcrumbItem>
+            </BreadcrumbList>
+          </Breadcrumb>
+        </div>
       </header>
 
-      <main className="space-y-8">
-        {/* Floating Trial Timer for Trial Users Only */}
-        {isTrial && trialTimer && (
-          <FloatingTrialTimer timer={trialTimer} />
-        )}
+      {/* Mobile Hub View */}
+      <div className="block sm:hidden">
+        <MobileDashboardHub />
+      </div>
 
-        {/* Activity Timers - Hidden for everyone (or specifically Trial users if logic dictates) */}
-        {/*
-          // COMMENTED OUT AS PER REQUEST to hide the checklist component
-          // Logic: We might want to show allowed timers for non-trial users if any exist?
-          // But for now, specifically hiding the Trial Checklist for users with membership tier (non-trial).
-          // And generally commenting out the component showing their currently active timer/checklist.
-        */}
-        {/* {timers && timers.length > 0 && !isTrial && (
-           <div className="space-y-6">
-             {timers.map((timer) => (
-               <ActivityTimerCard
-                 key={timer.id}
-                 timer={timer}
-                 onPause={() => pauseTimerMutation.mutate()}
-                 onResume={() => resumeTimerMutation.mutate()}
-                 isActionPending={pauseTimerMutation.isPending || resumeTimerMutation.isPending}
-               />
-             ))}
-           </div>
-        )} */}
-        {/* Completely commented out for now as requested "commnet our the component showing their currently" */}
-        {/* 
-        {timers && timers.length > 0 && (
-          <div className="space-y-6">
-            {timers.map((timer) => (
-              <ActivityTimerCard
-                key={timer.id}
-                timer={timer}
-                onPause={() => pauseTimerMutation.mutate()}
-                onResume={() => resumeTimerMutation.mutate()}
-                isActionPending={pauseTimerMutation.isPending || resumeTimerMutation.isPending}
-              />
-            ))}
-          </div>
-        )} 
-        */}
-
+      {/* Desktop Overview View */}
+      <main className="hidden sm:block space-y-8">
         {isLoadingStats ? (
           <p>Loading statistics...</p>
         ) : stats && userRole ? (
@@ -208,13 +131,9 @@ const DashboardPage: FC = () => {
                   isLoading={isLoadingActivities}
                 />
               </div>
-              {userRole === UserRole.OWNER && (
+              {userRole === 'owner' && (
                 <div className="lg:col-span-2 space-y-8">
-                  <ListingPackages
-                    name={packageInfo?.planType || 'Basic'}
-                    current={currentListings}
-                    max={maxListings}
-                  />
+                  <ListingPackages pkg={listingPackage} />
                   <EarningProgressionChart />
                 </div>
               )}
